[타워 강화 기능 구현 완료]

+++ Assets/Scripts/TowerData.cs
@@ -10,7 +10,7 @@ public class Tower
 
     public int money;
 
-    public int MD;
+    public float MD;
 
     public float range; 
 }
